third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-into-across-timeouts.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    setTimeout(callback1, 0);
}

function callback1()
{
    setTimeout(callback2, 0);
    debugger;
}

function callback2()
{
    var dummy = 42; // Should pause here.
    (function FAIL_Should_Not_Pause_Here() { debugger; })();

    if (window.testRunner)
        testRunner.evaluateInWebInspector(0, "(InspectorTest.completeDebuggerTest())")
}

var test = function()
{
    InspectorTest.startDebuggerTest(step1, true);

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2()
    {
        var actions = [
            "Print", // "debugger" in callback1
            "StepInto",
            "StepInto",
            "Print",
        ];
        InspectorTest.waitUntilPausedAndPerformSteppingActions(actions, resume);
    }

    function resume()
    {
        InspectorTest.resumeExecution(InspectorTest.waitUntilPaused.bind(InspectorTest, resume));
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that debugger StepInto will stop inside next timeout handler.
</p>
</body>
</html>
