third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-into-across-timeouts-expected.txt

Tests that debugger StepInto will stop inside next timeout handler.

Set timer for test function.
Call stack:
    0) callback1 (debugger-step-into-across-timeouts.html:15)

Executing StepInto...
Executing StepInto...
Call stack:
    0) callback2 (debugger-step-into-across-timeouts.html:18)
